requirements.txt
jupyter
PyGame
PyOpenGL
